More convenient wait_for_any macros
