+++ crates/holochain/src/conductor/interface/websocket.rs
@@ -707,7 +707,7 @@ pub mod test {
         let p2p_store = AgentKv::new(env.clone().into()).unwrap();
 
         // - Give time for the agents to join the network.
-        crate::wait_for_any_10s!(
+        crate::assert_eq_retry_10s!(
             {
                 fresh_reader_test!(env, |r| p2p_store
                     .as_store_ref()
@@ -716,8 +716,7 @@ pub mod test {
                     .count()
                     .unwrap())
             },
-            |&count| count == 4,
-            |count| assert_eq!(count, 4)
+            4
         );
 
         // - Get agents and space
